refactor(#51): GeocodingService 클래스명 수정 및 응답 구조를 static record로 변경
