+++ src/main/java/com/shiftm/shiftm/infra/location/KakaoGeocodingClient.java
@@ -1,7 +1,8 @@
-package com.shiftm.shiftm.infra.geocoding;
+package com.shiftm.shiftm.infra.location;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.HttpEntity;
@@ -10,8 +11,11 @@ import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
+import java.util.List;
+
+@Slf4j
 @Component
-public class GeocodingService {
+public class KakaoGeocodingClient {
 
     @Value("${kakao.api.key}")
     private String kakaoApiKey;
@@ -21,7 +25,7 @@ public class GeocodingService {
     private final RestTemplate restTemplate;
     private final ObjectMapper objectMapper;
 
-    public GeocodingService(RestTemplateBuilder restTemplateBuilder, ObjectMapper objectMapper) {
+    public KakaoGeocodingClient(RestTemplateBuilder restTemplateBuilder, ObjectMapper objectMapper) {
         this.restTemplate = restTemplateBuilder.build();
         this.objectMapper = objectMapper;
     }
@@ -33,15 +37,19 @@ public class GeocodingService {
 
         final String response = restTemplate.exchange(GEOCODING_API_URL, HttpMethod.GET, entity, String.class, longitude, latitude).getBody();
         try {
-            final GeocodingResponse geocodingResponse = objectMapper.readValue(response, GeocodingResponse.class);
+            final KakaoGeocodingResponse geocodingResponse = objectMapper.readValue(response, KakaoGeocodingResponse.class);
             if (geocodingResponse.documents() != null && !geocodingResponse.documents().isEmpty()) {
-                final String address = geocodingResponse.documents().get(0).address().addressName();
-                return address;
+                return geocodingResponse.documents().get(0).address().addressName();
             }
         } catch (Exception e) {
-            e.printStackTrace();
+            log.error("주소 변환 실패: ", e);
             return "주소 변환 실패";
         }
         return "주소를 찾을 수 없습니다";
+    }
+
+    public record KakaoGeocodingResponse(List<Document> documents) {
+        public record Document(@JsonProperty("address") Address address) {}
+        public record Address(@JsonProperty("address_name") String addressName) {}
     }
 }
